closing ticket# 65: add loading skeletons to AdminConsole page for devices loading enhancement v2, skeleton in place and working, UI enhancement refining of skeleton is optional
